refactor(aio): select federated auth plugins by plugin code (sync parity)
Review feedback on #1257 (idp_registry.py:74): the federated_auth code
now binds the ADFS implementation directly and only VALIDATES IDP_NAME
(empty or 'adfs' accepted, anything else raises UnsupportedIdp),
mirroring sync's FederatedAuthPluginFactory (federated_plugin.py:182-192);
the okta code binds the Okta implementation. The dynamic
AsyncIdpPluginRegistry and its IDP_NAME dispatch are removed -- plugin
codes are the selection mechanism, and custom-IdP extensibility can
follow the sync register_plugin discussion upstream.
